requirements.txt
streamlit==1.26.0
python-dotenv==1.0.0
langchain==0.0.275
langchain-google-genai==0.0.3
langchain-community==0.0.3
faiss-cpu==1.7.4